pyproject.toml
[tool.poetry]
name = "customer-support-backend"
version = "0.1.0"
description = "A FastAPI-based backend for a customer support assistant"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{ include = "app", from = "." }]


[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = ">=0.115.0"
uvicorn = ">=0.30.6"
sqlalchemy = ">=2.0.35"
pydantic = ">=2.9.2"
python-jose = ">=3.3.0"
passlib = ">=1.7.4"
pydantic-settings = ">=2.5.2"
python-multipart = ">=0.0.12"
aiohttp = ">=3.10.5"
psycopg2-binary = ">=2.9.9"
groq = ">=0.11.0"
python-dotenv = ">=1.0.1"
alembic = ">=1.13.3"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3.3"
pytest-asyncio = ">=0.24.0"
black = ">=24.8.0"
isort = ">=5.13.2"
mypy = ">=1.11.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"